Move the graph structs to model
